pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "multiai"
version = "0.1.0"
description = "Multi-AI Orchestrator with governed dataops pipeline"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "click>=8",
  "rich>=13",
  "pydantic>=2",
  "pyyaml>=6",
  "pandas>=2.2",
  "pyarrow>=16",
  "polars>=1.0",
  "httpx>=0.27",
  "websockets>=13",
  "orjson>=3.10",
]

[project.scripts]
multiai = "multiai.cli.main:main"
